Add atomicity and locking to object actions.
By default, ``ObjectAction`` and ``ObjectsAction`` lock the object(s)
being updated with a select_for_update and wrap the action in a transaction.
Since this is done with the ``djarg.qset`` helper, transactions and locking
will *not* happen when only running validations.

This behavior was documented and can be turned off by setting the
``select_for_update`` attribute of actions to ``None``

Type: feature