Don't override action methods on viewset classes
When DAF overrides the class definition with actions, it now will check
if the action is already defined and ignore setting the class attribute.
Previously it asserted it wasn't there and failed loudly, but that caused
issues for subclasses. Overriding it also causes issues for subclasses.

For now, we are ignoring adding the action if it is already defined.
We may want to revisit this behavior in the future.

Type: bug